requirements.txt
PyQt5
pytesseract
Pillow
opencv-python
scikit-image
requests